feat(resources): define a resource file
Resource are often the sane across your pipelines, therefore it would be nice to reduce duplication and define them in a single file. However not every resource should be added to every pipelines. The `resources_file` or `-r` flag let you declare a file to be merge in your pipeline, but unused *resources* and *resource_types* will be removed.